New interrupts & out of bounds management
Created a whole bunch of new interrupts to handle out of bounds, sys
calls (which according to the PDF are software interrupts?), and a slew
of tiny changes based on testing a bunch of programs. I should've split
the commits up a bit here.

For memory, got rid of looping address. Too difficult to deal with
updating the stupid RAM display to bother. If memory tries to go out of
bounds, it fires interrupt. Also modified the load and run program in
shell to be a bit more informative.