Chrome Bugfix???
I debug primarily in IE. Apparently, programs weren't parsing in Chrome
properly due to the innerHTML property of the program box?! I have no
idea if this was how it was when I submitted P1. I recall trying it in
Chrome and it worked. But debugging it today doesnt work as innerHTML
was blank....