RunAll & Basic CPUScheduler
Run all programs loaded consecutively. The runall queues up all the PCBs
in the residentq to the readyq, but if there is a program already
executing (eg after the 1st program gets queued it should immediately
run, so this applies to the 2nd program), it will tell the CPU scheduler
there are more programs waiting to be executed. Then when the CPU is
done executing the first program, it will fire an IRQ for BRK. In BRK,
it checks with the scheduler to see if there are programs waiting to be
run. If so, then scheduler will fire another IRQ to run the next
program. Repeat until ready queue is empty. The one issue I had was that
need Global PCB for the current running PCB, due to the fact that now,
after BRK, we throw away with the PCB from the readyQ. This breaks the
updating tables, which used to peak at the PCB that was running in the
queue. There will need to be adjustments for the rest of P3 to work
here. /longUpdate.