Priority and FCFS scheduling
I'm pretty sure this is all that's needed. Will email Alan to be sure.
Priority scheduling just sorts the Ready Queue whenever a process gets
added. Tested even adding a new process while it runs. The CPU needed to
fix its dequeue so it absolutely dequeues the running process (not the
first in the queue) when doing Priority. But thats all. And FCFS is
just... first in the queue which is standard by definition.