Fixed Swapping Bugs - P4 Done
There was a major issue with context switch timing and memory swapping.
Firstly, context switching could occur when I was preemptively trying to
swap EXEs. Weird timing issue took a long time to debug. The second
issue may or may not be critical depending on the design. If a program
finishes and there are still swap programs waiting to be executed, but
are not immediately next in the ready queue, should they be aware of the
newly available memory or should they only swap with the next TSB? E.G -
P0 and P1 in memory. P0 finishes. P1 finishes as normal, but P3 is in
page. P3 swaps into P2 memory, but then what about P4 in page? Does it
swap into P2's and P3's memory or should it swap into P1's older memory?
I opted for that as it felt right. The OS should be using all the
available RAM if there are EXEs in page at all times. These were the
hardest fixes the entire project. Thus, I am now finished.